Code from main moved to board
Also created an enum class and organised files into
separate 'bugs' and 'game' directories
