+++ C++ CA2/bugs/Bug.cpp
