+++ C++ CA2/bugs/Crawler.h
@@ -4,10 +4,11 @@
 #ifndef LUKE_HILLIARD_CPP_BUGS_LIFE_CRAWLER_H
 #define LUKE_HILLIARD_CPP_BUGS_LIFE_CRAWLER_H
 #include "Bug.h"
+#include <string>
 
 
 
-class Crawler : public  Bug {
+class Crawler : public Bug {
 
 public:
     Crawler(int id, pair<int, int> position, Direction direction, int size, bool alive, list<pair<int, int> > path  ) {
@@ -20,6 +21,20 @@ public:
     }
 
     Crawler() {}
+
+    string toString() {
+        string status;
+        if(alive)
+            status = "Alive";
+        else
+            status = "Dead";
+        return "| Crawler (" + to_string(id) + ") | " +
+                 "Position: (" + to_string(position.first) + ", " + to_string(position.second) + ") |  " +
+                 "Status: " + status + " | "
+                 "Size  : " + to_string(size) + " | " +
+                 "Facing: " + to_string(direction) + " |\n";
+
+    }
 
     virtual void move() override;
 };
