+++ C++ CA2/bugs/Direction.h
@@ -1,0 +1,12 @@
+//
+// Created by Luke Hilliard on 09/04/2024.
+//
+
+#ifndef C___CA2_DIRECTION_H
+#define C___CA2_DIRECTION_H
+#include <iostream>
+#include <map>
+
+enum class Direction {NORTH = 1, EAST = 2, SOUTH = 3, WEST = 4};
+
+#endif //C___CA2_DIRECTION_H
