+++ C++ CA2/bugs/Hopper.cpp
