+++ C++ CA2/bugs/Hopper.h
@@ -4,7 +4,7 @@
 #ifndef C___CA2_HOPPER_H
 #define C___CA2_HOPPER_H
 #include "Bug.h"
-
+#include <string>
 
 
 class Hopper:public Bug {
@@ -24,6 +24,20 @@ public :
     Hopper() {}
 
     virtual void move() override;
+
+    string toString() {
+        string status;
+        if(alive)
+            status = "Alive";
+        else
+            status = "Dead";
+        return "| Hopper (" + to_string(id) + ") | Hop Length: " + to_string(hopLength) + " | "
+               "Position: (" + to_string(position.first) + ", " + to_string(position.second) + ") |  " +
+               "Status: " + status + " | "
+               "Size  : " + to_string(size) + " | " +
+               "Facing: " + to_string(direction) + " |\n";
+
+    }
 
 };
