+++ C++ CA2/cmake-build-debug/.cmake/api/v1/reply/target-C___CA2-Debug-338fc5cfbd620f705c31.json
@@ -52,7 +52,8 @@
 				0,
 				3,
 				4,
-				6
+				6,
+				7
 			]
 		}
 	],
@@ -88,7 +89,8 @@
 				0,
 				3,
 				4,
-				6
+				6,
+				7
 			]
 		},
 		{
@@ -97,7 +99,9 @@
 			[
 				1,
 				2,
-				5
+				5,
+				8,
+				9
 			]
 		}
 	],
@@ -111,36 +115,52 @@
 		},
 		{
 			"backtrace" : 1,
-			"path" : "Bug.h",
+			"path" : "bugs/Bug.h",
 			"sourceGroupIndex" : 1
 		},
 		{
 			"backtrace" : 1,
-			"path" : "Crawler.h",
-			"sourceGroupIndex" : 1
-		},
-		{
-			"backtrace" : 1,
-			"compileGroupIndex" : 0,
-			"path" : "Crawler.cpp",
-			"sourceGroupIndex" : 0
-		},
-		{
-			"backtrace" : 1,
-			"compileGroupIndex" : 0,
-			"path" : "Bug.cpp",
-			"sourceGroupIndex" : 0
-		},
-		{
-			"backtrace" : 1,
-			"path" : "Hopper.h",
+			"path" : "bugs/Crawler.h",
 			"sourceGroupIndex" : 1
 		},
 		{
 			"backtrace" : 1,
 			"compileGroupIndex" : 0,
-			"path" : "Hopper.cpp",
+			"path" : "bugs/Crawler.cpp",
 			"sourceGroupIndex" : 0
+		},
+		{
+			"backtrace" : 1,
+			"compileGroupIndex" : 0,
+			"path" : "bugs/Bug.cpp",
+			"sourceGroupIndex" : 0
+		},
+		{
+			"backtrace" : 1,
+			"path" : "bugs/Hopper.h",
+			"sourceGroupIndex" : 1
+		},
+		{
+			"backtrace" : 1,
+			"compileGroupIndex" : 0,
+			"path" : "bugs/Hopper.cpp",
+			"sourceGroupIndex" : 0
+		},
+		{
+			"backtrace" : 1,
+			"compileGroupIndex" : 0,
+			"path" : "game/Board.cpp",
+			"sourceGroupIndex" : 0
+		},
+		{
+			"backtrace" : 1,
+			"path" : "game/Board.h",
+			"sourceGroupIndex" : 1
+		},
+		{
+			"backtrace" : 1,
+			"path" : "bugs/Direction.h",
+			"sourceGroupIndex" : 1
 		}
 	],
 	"type" : "EXECUTABLE"
