+++ C++ CA2/cmake-build-debug/CMakeFiles/C___CA2.dir/cmake_clean.cmake
@@ -1,10 +1,12 @@
 file(REMOVE_RECURSE
-  "CMakeFiles/C___CA2.dir/Bug.cpp.o"
-  "CMakeFiles/C___CA2.dir/Bug.cpp.o.d"
-  "CMakeFiles/C___CA2.dir/Crawler.cpp.o"
-  "CMakeFiles/C___CA2.dir/Crawler.cpp.o.d"
-  "CMakeFiles/C___CA2.dir/Hopper.cpp.o"
-  "CMakeFiles/C___CA2.dir/Hopper.cpp.o.d"
+  "CMakeFiles/C___CA2.dir/bugs/Bug.cpp.o"
+  "CMakeFiles/C___CA2.dir/bugs/Bug.cpp.o.d"
+  "CMakeFiles/C___CA2.dir/bugs/Crawler.cpp.o"
+  "CMakeFiles/C___CA2.dir/bugs/Crawler.cpp.o.d"
+  "CMakeFiles/C___CA2.dir/bugs/Hopper.cpp.o"
+  "CMakeFiles/C___CA2.dir/bugs/Hopper.cpp.o.d"
+  "CMakeFiles/C___CA2.dir/game/Board.cpp.o"
+  "CMakeFiles/C___CA2.dir/game/Board.cpp.o.d"
   "CMakeFiles/C___CA2.dir/main.cpp.o"
   "CMakeFiles/C___CA2.dir/main.cpp.o.d"
   "C___CA2"
