+++ C++ CA2/cmake-build-debug/CMakeFiles/C___CA2.dir/link.txt
@@ -1,1 +1,1 @@
-/Library/Developer/CommandLineTools/usr/bin/c++ -g -arch arm64 -isysroot /Library/Developer/CommandLineTools/SDKs/MacOSX14.2.sdk -Wl,-search_paths_first -Wl,-headerpad_max_install_names CMakeFiles/C___CA2.dir/main.cpp.o CMakeFiles/C___CA2.dir/Crawler.cpp.o CMakeFiles/C___CA2.dir/Bug.cpp.o CMakeFiles/C___CA2.dir/Hopper.cpp.o -o C___CA2 
+/Library/Developer/CommandLineTools/usr/bin/c++ -g -arch arm64 -isysroot /Library/Developer/CommandLineTools/SDKs/MacOSX14.2.sdk -Wl,-search_paths_first -Wl,-headerpad_max_install_names CMakeFiles/C___CA2.dir/main.cpp.o CMakeFiles/C___CA2.dir/bugs/Crawler.cpp.o CMakeFiles/C___CA2.dir/bugs/Bug.cpp.o CMakeFiles/C___CA2.dir/bugs/Hopper.cpp.o CMakeFiles/C___CA2.dir/game/Board.cpp.o -o C___CA2
