+++ C++ CA2/cmake-build-debug/CMakeFiles/Makefile2
@@ -81,12 +81,12 @@ clean: CMakeFiles/C___CA2.dir/clean
 CMakeFiles/C___CA2.dir/all:
 	$(MAKE) $(MAKESILENT) -f CMakeFiles/C___CA2.dir/build.make CMakeFiles/C___CA2.dir/depend
 	$(MAKE) $(MAKESILENT) -f CMakeFiles/C___CA2.dir/build.make CMakeFiles/C___CA2.dir/build
-	@$(CMAKE_COMMAND) -E cmake_echo_color "--switch=$(COLOR)" --progress-dir="/Users/lukehilliard/Desktop/C++ CA2/luke-hilliard-cpp-bugs-life/C++ CA2/cmake-build-debug/CMakeFiles" --progress-num=1,2,3,4,5 "Built target C___CA2"
+	@$(CMAKE_COMMAND) -E cmake_echo_color "--switch=$(COLOR)" --progress-dir="/Users/lukehilliard/Desktop/C++ CA2/luke-hilliard-cpp-bugs-life/C++ CA2/cmake-build-debug/CMakeFiles" --progress-num=1,2,3,4,5,6 "Built target C___CA2"
 .PHONY : CMakeFiles/C___CA2.dir/all
 
 # Build rule for subdir invocation for target.
 CMakeFiles/C___CA2.dir/rule: cmake_check_build_system
-	$(CMAKE_COMMAND) -E cmake_progress_start "/Users/lukehilliard/Desktop/C++ CA2/luke-hilliard-cpp-bugs-life/C++ CA2/cmake-build-debug/CMakeFiles" 5
+	$(CMAKE_COMMAND) -E cmake_progress_start "/Users/lukehilliard/Desktop/C++ CA2/luke-hilliard-cpp-bugs-life/C++ CA2/cmake-build-debug/CMakeFiles" 6
 	$(MAKE) $(MAKESILENT) -f CMakeFiles/Makefile2 CMakeFiles/C___CA2.dir/all
 	$(CMAKE_COMMAND) -E cmake_progress_start "/Users/lukehilliard/Desktop/C++ CA2/luke-hilliard-cpp-bugs-life/C++ CA2/cmake-build-debug/CMakeFiles" 0
 .PHONY : CMakeFiles/C___CA2.dir/rule
