+++ C++ CA2/cmake-build-debug/Testing/Temporary/LastTest.log
@@ -1,3 +1,3 @@
-Start testing: Apr 09 16:44 IST
+Start testing: Apr 09 18:01 IST
 ----------------------------------------------------------
-End testing: Apr 09 16:44 IST
+End testing: Apr 09 18:01 IST
